+++ src/Symfony/Component/Console/Helper/DebugFormatterHelper.php
@@ -11,8 +11,6 @@
 
 namespace Symfony\Component\Console\Helper;
 
-use Symfony\Component\Console\Helper\Helper;
-
 /**
  * Helps outputting debug information when running an external program from a command.
  *
@@ -26,6 +24,15 @@ class DebugFormatterHelper extends Helper
     private $started = array();
     private $count = -1;
 
+    /**
+     * Starts a debug formatting session
+     *
+     * @param string $id      The id of the formatting session
+     * @param string $message The message to display
+     * @param string $prefix  The prefix to use
+     *
+     * @return string
+     */
     public function start($id, $message, $prefix = 'RUN')
     {
         $this->started[$id] = array('border' => ++$this->count % count($this->colors));
@@ -33,20 +40,39 @@ class DebugFormatterHelper extends Helper
         return sprintf("%s<bg=blue;fg=white> %s </> <fg=blue>%s</>\n", $this->getBorder($id), $prefix, $message);
     }
 
+    /**
+     * Adds progress to a formatting session
+     *
+     * @param string $id          The id of the formatting session
+     * @param string $buffer      The message to display
+     * @param bool   $error       Whether to consider the buffer as error
+     * @param string $prefix      The prefix for output
+     * @param string $errorPrefix The prefix for error output
+     *
+     * @return string
+     */
     public function progress($id, $buffer, $error = false, $prefix = 'OUT', $errorPrefix = 'ERR')
     {
         $message = '';
 
         if ($error) {
+            if (isset($this->started[$id]['out'])) {
+                $message .= "\n";
+                unset($this->started[$id]['out']);
+            }
             if (!isset($this->started[$id]['err'])) {
-                $message = sprintf("%s<bg=red;fg=white> %s </> ", $this->getBorder($id), $errorPrefix);
+                $message .= sprintf("%s<bg=red;fg=white> %s </> ", $this->getBorder($id), $errorPrefix);
                 $this->started[$id]['err'] = true;
             }
 
             $message .= str_replace("\n", sprintf("\n%s<bg=red;fg=white> %s </> ", $this->getBorder($id), $errorPrefix), $buffer);
         } else {
+            if (isset($this->started[$id]['err'])) {
+                $message .= "\n";
+                unset($this->started[$id]['err']);
+            }
             if (!isset($this->started[$id]['out'])) {
-                $message = sprintf("%s<bg=green;fg=white> %s </> ", $this->getBorder($id), $prefix);
+                $message .= sprintf("%s<bg=green;fg=white> %s </> ", $this->getBorder($id), $prefix);
                 $this->started[$id]['out'] = true;
             }
 
@@ -56,6 +82,16 @@ class DebugFormatterHelper extends Helper
         return $message;
     }
 
+    /**
+     * Stops a formatting session
+     *
+     * @param string $id         The id of the formatting session
+     * @param string $message    The message to display
+     * @param bool   $successful Whether to consider the result as success
+     * @param string $prefix     The prefix for the end output
+     *
+     * @return string
+     */
     public function stop($id, $message, $successful, $prefix = 'RES')
     {
         $trailingEOL = isset($this->started[$id]['out']) || isset($this->started[$id]['err']) ? "\n" : '';
@@ -64,16 +100,25 @@ class DebugFormatterHelper extends Helper
             return sprintf("%s%s<bg=green;fg=white> %s </> <fg=green>%s</>\n", $trailingEOL, $this->getBorder($id), $prefix, $message);
         }
 
-        return sprintf("%s%s<bg=red;fg=white> %s </> <fg=red>%s</>\n", $trailingEOL, $this->getBorder($id), $prefix, $message);
+        $message = sprintf("%s%s<bg=red;fg=white> %s </> <fg=red>%s</>\n", $trailingEOL, $this->getBorder($id), $prefix, $message);
+
+        unset($this->started[$id]['out'], $this->started[$id]['err']);
+
+        return $message;
     }
 
+    /**
+     * @param string $id The id of the formatting session
+     *
+     * @return string
+     */
     private function getBorder($id)
     {
         return sprintf('<bg=%s> </>', $this->colors[$this->started[$id]['border']]);
     }
 
     /**
-     * {@inheritDoc}
+     * {@inheritdoc}
      */
     public function getName()
     {